updated uninstall.ps1

+++ pwsh/uninstall.ps1
@@ -490,10 +490,10 @@ foreach ($app in $selectedApps) {
             Write-Host "Uninstalling Hot Corners..." -ForegroundColor Yellow
             Stop-Process -n WinXCorners -Force
             Stop-Process -n WinLaunch -Force
-            Stop-Process -n ssn -Force
-            Remove-Item -Path "$env:LOCALAPPDATA\WinXCorners" -Recurse -Force
+            Stop-Process -n ssn -Force            
             Remove-Item -Path "$env:LOCALAPPDATA\WinMac\hotcorners" -Recurse -Force
-            if (-not (Get-Process -Name WinLaunch -ErrorAction SilentlyContinue)) { Remove-Item -Path "$env:LOCALAPPDATA\WinLaunch" -Recurse -Force }
+            if (-not (Get-Process -Name WinLaunch -ErrorAction SilentlyContinue)) { Remove-Item -Path "$env:LOCALAPPDATA\WinXCorners" -Recurse -Force }
+            if (-not (Get-Process -Name WinLaunch -ErrorAction SilentlyContinue)) { Remove-Item -Path "$env:LOCALAPPDATA\WinLaunch" -Recurse -Force; Remove-Item -Path "$env:APPDATA\WinLaunch" -Recurse -Force }
             Invoke-Output { Uninstall-WinGetPackage -name "Simple Sticky Notes" }
             Invoke-Output { Install-WinGetPackage 9NBLGGH4QGHW }
             $regPath1 = "HKCU:\Software\Microsoft\Windows\CurrentVersion\Run"
